Move clone components
