+++ ui/src/KaptureApp/components/GridTabsConnected.js
@@ -2,11 +2,8 @@ import { connect } from 'react-redux';
 
 import { GridTabs } from 'AtlasUI/components';
 import { editor } from 'KaptureApp/store';
-import { actions } from 'KaptureApp/actions';
 import { COMPONENT_TYPES } from 'KaptureApp/config/componentTypes';
 import { CONTAINER_TYPE_OPTIONS } from 'KaptureApp/config/containerTypes';
-
-const { cloneGrid } = actions.editor;
 
 const mapState = (state, props) => {
   return {
@@ -22,7 +19,7 @@ const mapDispatch = {
   onAddPlates: editor.addNewPlates,
   onAddRack: editor.addNewRack,
   onAddContainer: editor.addNewContainer,
-  onClone: cloneGrid,
+  onClone: editor.cloneGrid,
   onDelete: editor.deleteGrid,
 };
